+++ assets/src/components/component/ComponentTree.tsx
@@ -1,23 +1,15 @@
 import { EmptyState } from '@pluralsh/design-system'
 import { useMemo } from 'react'
 import { useOutletContext } from 'react-router-dom'
-
-import { ComponentTreeFragment, useComponentTreeQuery } from 'generated/graphql'
-
-import { GqlError } from 'components/utils/Alert'
-
-import { isNonNullable } from 'utils/isNonNullable'
-
-import { ComponentTreeFlow } from 'components/cd/pipelines/ComponentTreeFlow'
-
 import { ReactFlowProvider } from 'reactflow'
 
+import { useComponentTreeQuery } from 'generated/graphql'
+
+import { GqlError } from 'components/utils/Alert'
+import { ComponentTreeFlow } from 'components/component/tree/ComponentTreeFlow'
+
 import { ComponentDetailsContext } from './ComponentDetails'
-import {
-  C_TYPES,
-  HasMetadata,
-  getTreeNodesAndEdges,
-} from './tree/getTreeNodesAndEdges'
+import { getTreeNodesAndEdges } from './tree/getTreeNodesAndEdges'
 
 export default function ComponentTree() {
   const ctx = useOutletContext<ComponentDetailsContext>()
@@ -28,14 +20,7 @@ export default function ComponentTree() {
 
   console.log('queryRes', queryRes)
 
-  const flowData = useMemo(() => {
-    const metadatas = flattenMetadata(tree)
-
-    return getTreeNodesAndEdges({
-      edges: tree?.edges?.filter?.(isNonNullable) || [],
-      metadatas: metadatas || [],
-    })
-  }, [tree])
+  const flowData = useMemo(() => getTreeNodesAndEdges(tree), [tree])
 
   console.log('flow', flowData)
 
@@ -46,37 +31,23 @@ export default function ComponentTree() {
     return <EmptyState message="No data available." />
   }
 
-  // return (
-  //   <div>
-  //     edges:
-  //     {queryRes.data.componentTree.edges?.map((e) => (
-  //       <div>
-  //         {e?.from}--{e?.to}
-  //       </div>
-  //     ))}
-  //   </div>
-  // )
-
   return (
-    <ReactFlowProvider>
-      <ComponentTreeFlow
-        nodes={flowData.nodes}
-        edges={flowData.edges}
-      />
-    </ReactFlowProvider>
+    <div css={{ height: '100%', display: 'flex', flexDirection: 'column' }}>
+      <div>
+        <div>Edges</div>
+        {queryRes.data.componentTree.edges?.map((e) => (
+          <div>
+            Edge: {e?.from}--{e?.to}
+          </div>
+        ))}
+      </div>
+
+      <ReactFlowProvider>
+        <ComponentTreeFlow
+          nodes={flowData.nodes}
+          edges={flowData.edges}
+        />
+      </ReactFlowProvider>
+    </div>
   )
 }
-
-function flattenMetadata(tree: Nullable<ComponentTreeFragment>) {
-  if (!tree) return []
-
-  return C_TYPES.flatMap(
-    (cType) =>
-      tree?.[cType]?.filter(isNonNullable).map((c: HasMetadata) => ({
-        kind: cType.slice(0, -1),
-        ...c,
-      })) || []
-  )
-}
-
-export type TreeNodeMeta = ReturnType<typeof flattenMetadata>[number]
