+++ assets/src/components/component/tree/getTreeNodesAndEdges.tsx
@@ -1,8 +1,10 @@
 import { ComponentTreeFragment, MetadataFragment } from 'generated/graphql'
 import { ConditionalKeys } from 'type-fest'
+import { isNonNullable } from 'utils/isNonNullable'
 
-import { TreeNodeMeta } from '../ComponentTree'
+import { baseEdgeProps } from 'components/cd/pipelines/utils/getNodesAndEdges'
 
+export type TreeNodeMeta = ReturnType<typeof flattenMetadata>[number]
 export type HasMetadata = { metadata?: MetadataFragment }
 export type ComponentKindsKey = ConditionalKeys<
   ComponentTreeFragment,
@@ -21,34 +23,36 @@ export const C_TYPES = [
   'statefulsets',
 ] as const satisfies ComponentKindsKey[]
 
-// import { baseEdgeProps } from 'components/cd/pipelines/utils/getNodesAndEdges'
-const baseEdgeProps = {
-  type: 'plural-edge',
-  updatable: false,
-}
 const TYPE_SORT_VALS = Object.fromEntries(
   C_TYPES.map((val, i) => [val, i])
 ) as Record<ComponentKindsKey, number>
 
-export function getTreeNodesAndEdges({
-  edges: edgesProp,
-  metadatas,
-}: {
-  edges: { from: string; to: string }[]
-  metadatas: TreeNodeMeta[]
-}) {
-  const edges = edgesProp.flatMap((edge) => {
-    if (!edge) return []
+function flattenMetadata(tree: Nullable<ComponentTreeFragment>) {
+  if (!tree) return []
 
-    return {
-      ...baseEdgeProps,
-      id: `${edge.from}->${edge.to}`,
-      source: edge.from,
-      target: edge.to,
-    }
-  })
+  return C_TYPES.flatMap(
+    (cType) =>
+      tree?.[cType]?.filter(isNonNullable).map((c: HasMetadata) => ({
+        kind: cType.slice(0, -1),
+        ...c,
+      })) || []
+  )
+}
 
-  const nodes = metadatas
+export function getTreeNodesAndEdges(tree: Nullable<ComponentTreeFragment>) {
+  const edges =
+    tree?.edges?.flatMap((edge) => {
+      if (!edge) return []
+
+      return {
+        ...baseEdgeProps,
+        id: `${edge.from}->${edge.to}`,
+        source: edge.from,
+        target: edge.to,
+      }
+    }) || []
+
+  const nodes = flattenMetadata(tree)
     // Order of edges matters to Dagre layout, so sort by type ahead of time
     .sort((a, b) => TYPE_SORT_VALS[b.kind] - TYPE_SORT_VALS[a.kind])
     .map(({ kind, metadata }) => {
